Added support for multi-dimensional arrays in the grammar, compiler and API. Fixed some issues regarding error row output. Added alternative spellings, for example groupshared can be expressed using snake case (group_shared). This applies to a bunch of multiple word attributes. Removed dead code. Replaced all occurrences of LP, RP etc with the actual character in the grammar, makes it easier to read. Tabs to spaces. Fixed incomplete error output from glslang. Actually, we can just print this output raw, since we provide #line directives, the errors should occur on the proper lines.

+++ anyfx/compiler/code/parser4/anyfxerrorhandlers.h
@@ -12,7 +12,7 @@ public:
 	bool hasError = false;
 	std::string errorBuffer;
 	std::string warningBuffer;
-	std::vector<std::tuple<int, size_t, size_t, size_t, std::string>> lines;
+	std::vector<std::tuple<int,int,int,int, std::string>> lines;
 };
 
 class AnyFXParserHandler : public antlr4::BaseErrorListener
@@ -24,5 +24,5 @@ public:
 	bool hasError = false;
 	std::string errorBuffer;
 	std::string warningBuffer;
-	std::vector<std::tuple<int, size_t, size_t, size_t, std::string>> lines;
+	std::vector<std::tuple<int,int,int,int, std::string>> lines;
 };
